refactor(console): app page structure
