+++ packages/console/src/containers/AppContent/Sidebar/components/Item/index.module.scss
