+++ packages/console/src/containers/AppContent/Sidebar/components/Section/index.module.scss
