+++ packages/console/src/containers/AppContent/Sidebar/icons/BarGraph.tsx
