+++ packages/console/src/containers/AppContent/Sidebar/icons/Bolt.tsx
