+++ packages/console/src/containers/AppContent/Sidebar/icons/Box.tsx
