+++ packages/console/src/containers/AppContent/Sidebar/icons/Cloud.tsx
