+++ packages/console/src/containers/AppContent/Sidebar/icons/Connection.tsx
