+++ packages/console/src/containers/AppContent/Sidebar/icons/Contact.tsx
