+++ packages/console/src/containers/AppContent/Sidebar/icons/Document.tsx
