+++ packages/console/src/containers/AppContent/Sidebar/icons/Gear.tsx
