+++ packages/console/src/containers/AppContent/Sidebar/icons/List.tsx
