+++ packages/console/src/containers/AppContent/Sidebar/icons/UserProfile.tsx
