+++ packages/console/src/containers/AppContent/Sidebar/icons/Web.tsx
