+++ packages/console/src/containers/AppContent/Sidebar/index.module.scss
